requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bleach==6.0.0
boto3==1.26.151
botocore==1.29.151
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-choices==1.7.2
django-cors-headers==4.0.0
django-extensions==3.2.3
django-filter==23.2
django-positions==0.6.0
django-rest-framework==0.1.0
django-solo==2.1.0
django-storages==1.13.2
django-webpack-loader==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
httplib2==0.22.0
idna==3.4
install==1.3.5
jmespath==1.0.1
multidict==6.0.4
nameparser==1.1.2
oauth2client==4.1.3
oauthlib==3.2.2
Pillow==9.5.0
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
sendgrid==6.10.0
sentry-sdk==1.26.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stripe==5.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.5.0
yarl==1.9.2
